Move the export action to the object list context menu.
